backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pydantic
pydantic-settings
python-dotenv
langchain
langchain-openai
openai
requests
python-multipart
pytest
